Abstracted CSV File loading function to CSVFileREader class
